Move getParameterChoices to fix imports.
