+++ newIDE/app/src/EventsSheet/ParameterFields/ParameterMetadataTools.js
@@ -70,3 +70,25 @@ export const getLastObjectParameterObjectType = (
 
   return parameters.at(objectParameterIndex).getExtraInfo();
 };
+
+export const getParameterChoices = (
+  parameterMetadata: ?gdParameterMetadata
+): Array<ExpressionAutocompletion> => {
+  if (!parameterMetadata) {
+    return [];
+  }
+
+  try {
+    return JSON.parse(parameterMetadata.getExtraInfo()).map(choice => ({
+      kind: 'Text',
+      completion: `"${choice}"`,
+    }));
+  } catch (exception) {
+    console.error(
+      'The parameter seems misconfigured, as an array of choices could not be extracted - verify that your properly wrote a list of choices in JSON format. Full exception is:',
+      exception
+    );
+  }
+
+  return [];
+};
